renamed kuzzle.js -> Kuzzle.js
