+++ src/Kuzzle.js
