+++ test/kuzzleDataMapping/constructor.test.js
@@ -2,7 +2,7 @@ var
   should = require('should'),
   rewire = require('rewire'),
   bluebird = require('bluebird'),
-  Kuzzle = rewire('../../src/kuzzle'),
+  Kuzzle = rewire('../../src/Kuzzle'),
   CollectionMapping = rewire('../../src/CollectionMapping');
 
 describe('CollectionMapping constructor', function () {
