+++ test/security/kuzzleSecurity/constructor.test.js
@@ -2,7 +2,7 @@ var
   should = require('should'),
   bluebird = require('bluebird'),
   rewire = require('rewire'),
-  Kuzzle = rewire('../../../src/kuzzle'),
+  Kuzzle = rewire('../../../src/Kuzzle'),
   Security = require('../../../src/security/kuzzleSecurity');
 
 describe('kuzzleSecurity constructor', function () {
