+++ test/security/kuzzleSecurity/rightsMethods.test.js
@@ -1,6 +1,6 @@
 var
   should = require('should'),
-  Kuzzle = require('../../../src/kuzzle');
+  Kuzzle = require('../../../src/Kuzzle');
 
 describe('Security user rights methods', function () {
   var
