+++ test/security/kuzzleSecurity/userMethods.test.js
@@ -1,6 +1,6 @@
 var
   should = require('should'),
-  Kuzzle = require('../../../src/kuzzle'),
+  Kuzzle = require('../../../src/Kuzzle'),
   User = require('../../../src/security/kuzzleUser');
 
 describe('Security user methods', function () {
