+++ test/security/kuzzleSecurityDocument/constructor.test.js
@@ -2,7 +2,7 @@ var
   should = require('should'),
   bluebird = require('bluebird'),
   rewire = require('rewire'),
-  Kuzzle = rewire('../../../src/kuzzle'),
+  Kuzzle = rewire('../../../src/Kuzzle'),
   KuzzleSecurityDocument = require('../../../src/security/kuzzleSecurityDocument');
 
 describe('KuzzleSecurityDocument constructor', function () {
