+++ test/security/kuzzleUser/methods.test.js
@@ -1,6 +1,6 @@
 var
   should = require('should'),
-  Kuzzle = require('../../../src/kuzzle'),
+  Kuzzle = require('../../../src/Kuzzle'),
   Role = require('../../../src/security/kuzzleRole'),
   User = require('../../../src/security/kuzzleUser');
